refactor(feedback toast): move toast from layout to template
in the previous commit, we added pathname to the toast's dependency
array to force it to show up even though it normally wouldn't because
layouts only render once. templates should fix this as they render
multiple times like pages. they are still a bit buggy though and at the
moment they don't work, so we keep pathname in the dependency array